further adaptations concerning TimeAllocationMutatorConfigGroup
git-svn-id: file:///srv/local/data/cleanRepo/trunk@19864 a0517364-e873-4167-b1c8-416daa5a6b91

+++ matsim/src/main/java/org/matsim/pt/replanning/TransitPlanMutateTimeAllocation.java
@@ -40,11 +40,11 @@ import org.matsim.pt.PtConstants;
  */
 public class TransitPlanMutateTimeAllocation implements PlanAlgorithm {
 
-	private final int mutationRange;
+	private final Double mutationRange;
 	private final Random random;
 	private boolean useActivityDurations;
 
-	public TransitPlanMutateTimeAllocation(final int mutationRange, final Random random) {
+	public TransitPlanMutateTimeAllocation(final Double mutationRange, final Random random) {
 		this.mutationRange = mutationRange;
 		this.random = random;
 	}
